moving component to own file
